feat: Implement storefront page layout customization, add RBAC system, and introduce newsletter subscription functionality.

+++ backend-laravel/routes/tenantModules/catalog.php
@@ -6,23 +6,23 @@ use App\Http\Controllers\Tenant\CategoriesController;
 use App\Http\Controllers\Tenant\BrandsController;
 
 // Products
-Route::get('/products', [ProductsController::class, 'index']);
-Route::get('/products/{id}', [ProductsController::class, 'show']);
-Route::post('/products', [ProductsController::class, 'store']);
-Route::put('/products/{id}', [ProductsController::class, 'update']);
-Route::delete('/products/{id}', [ProductsController::class, 'destroy']);
-Route::post('/products/{id}/adjust-stock', [ProductsController::class, 'adjustStock']);
+Route::get('/products', [ProductsController::class, 'index'])->middleware('permission:products.view');
+Route::get('/products/{id}', [ProductsController::class, 'show'])->middleware('permission:products.view');
+Route::post('/products', [ProductsController::class, 'store'])->middleware('permission:products.create');
+Route::put('/products/{id}', [ProductsController::class, 'update'])->middleware('permission:products.edit');
+Route::delete('/products/{id}', [ProductsController::class, 'destroy'])->middleware('permission:products.delete');
+Route::post('/products/{id}/adjust-stock', [ProductsController::class, 'adjustStock'])->middleware('permission:products.edit');
 
 // Categories
-Route::get('/categories', [CategoriesController::class, 'index']);
-Route::get('/categories/{id}', [CategoriesController::class, 'show']);
-Route::post('/categories', [CategoriesController::class, 'store']);
-Route::put('/categories/{id}', [CategoriesController::class, 'update']);
-Route::delete('/categories/{id}', [CategoriesController::class, 'destroy']);
+Route::get('/categories', [CategoriesController::class, 'index'])->middleware('permission:products.view');
+Route::get('/categories/{id}', [CategoriesController::class, 'show'])->middleware('permission:products.view');
+Route::post('/categories', [CategoriesController::class, 'store'])->middleware('permission:products.create');
+Route::put('/categories/{id}', [CategoriesController::class, 'update'])->middleware('permission:products.edit');
+Route::delete('/categories/{id}', [CategoriesController::class, 'destroy'])->middleware('permission:products.delete');
 
 // Brands
-Route::get('/brands', [BrandsController::class, 'index']);
-Route::get('/brands/{id}', [BrandsController::class, 'show']);
-Route::post('/brands', [BrandsController::class, 'store']);
-Route::put('/brands/{id}', [BrandsController::class, 'update']);
-Route::delete('/brands/{id}', [BrandsController::class, 'destroy']);
+Route::get('/brands', [BrandsController::class, 'index'])->middleware('permission:products.view');
+Route::get('/brands/{id}', [BrandsController::class, 'show'])->middleware('permission:products.view');
+Route::post('/brands', [BrandsController::class, 'store'])->middleware('permission:products.create');
+Route::put('/brands/{id}', [BrandsController::class, 'update'])->middleware('permission:products.edit');
+Route::delete('/brands/{id}', [BrandsController::class, 'destroy'])->middleware('permission:products.delete');
